ai-engine/requirements.txt
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
facenet-pytorch==2.5.3
opencv-python-headless==4.9.0.80
dlib==19.24.2
face-recognition==1.3.0
scikit-learn==1.3.2
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
numpy==1.26.3
Pillow==10.2.0
pytesseract==0.3.10
python-multipart==0.0.9
optuna==3.5.0
requests==2.31.0
pytest==7.4.4
httpx==0.26.0
